Fixes file uploading issue and updates user model for profile picture
Fixes an issue with file uploading in updateUser controller that was causing the response to always be successful, even when the file upload failed. Also updates the user model to include the "profilePic" field for storing the URL of the uploaded profile picture. This allows for a more accurate representation of the user's profile information.

Relates to #789