Adds "add to wishlist" functionality to anime details page
Enables users to add an anime to their wishlist from the anime details page. The "Add to List" button triggers the "handlewishlist" function, which makes a POST request to the server to add the anime to the user's wishlist. This feature enhances user experience by allowing them to easily create a personalized anime list.

Fixes #789